+++ src/components/layout/login/LoginUser.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent, FormEvent } from "react";
 import { useNavigate } from "react-router-dom";
 import { create } from "zustand";
-import NaverLogin from "../components/layout/login/NaverLogin";
-import KakaoLogin from "../components/layout/login/KakaoLogin";
+import NaverLogin from "./NaverLogin";
+import KakaoLogin from "./KakaoLogin";
 
 // Zustand store 정의
 interface LoginStore {
@@ -13,10 +13,10 @@ interface LoginStore {
 }
 
 const useLoginStore = create<LoginStore>((set) => ({
-  email: '',
-  password: '',
+  email: "",
+  password: "",
   setField: (field, value) => set((state) => ({ ...state, [field]: value })),
-  reset: () => set({ email: '', password: '' }),
+  reset: () => set({ email: "", password: "" }),
 }));
 
 // 가상의 로그인 함수 (실제로는 백엔드 API를 호출해야 합니다)
@@ -35,7 +35,7 @@ const LoginUser: React.FC = () => {
   const { email, password, setField, reset } = useLoginStore();
 
   const moveSignUpPage = () => {
-    navigate('/SignUp');  // 회원가입 페이지 경로
+    navigate("/SignUp"); // 회원가입 페이지 경로
   };
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -48,13 +48,13 @@ const LoginUser: React.FC = () => {
       alert("이메일과 비밀번호를 입력하세요.");
       return;
     }
-    
+
     try {
       const isValid = await checkCredentials(email, password);
       if (isValid) {
         console.log("로그인 성공");
         reset();
-        navigate('/');
+        navigate("/");
       } else {
         alert("이메일 또는 비밀번호가 올바르지 않습니다.");
       }
@@ -66,25 +66,25 @@ const LoginUser: React.FC = () => {
 
   return (
     <div className="flex flex-col justify-center items-center mt-24 mb-20 w-4/5">
-      <h1 className="text-3xl font-bold mb-12">LOGIN</h1>
+      <h1 className="text-3xl text-Blue-2 font-bold mb-12">LOGIN</h1>
       <form className="w-full flex flex-col justify-center items-center" onSubmit={submitLogin}>
         <div className="sm:w-96 flex justify-between items-center my-2">
-          <label className="w-16">이메일</label>
+          <label className="w-16 text-Blue-2">이메일</label>
           <input
             type="email"
             name="email"
-            className="w-40 sm:w-56 border-solid border-2 border-gray-200 rounded sm:ml-12 placeholder:text-slate-400 placeholder:text-xs indent-2.5 py-1"
+            className="input w-40 sm:w-56 sm:ml-12 placeholder:text-slate-400 placeholder:text-xs"
             placeholder="example@example.com"
             value={email}
             onChange={handleChange}
           />
         </div>
         <div className="sm:w-96 flex justify-between items-center mb-2">
-          <label className="w-16">비밀번호</label>
+          <label className="w-16 text-Blue-2">비밀번호</label>
           <input
             type="password"
             name="password"
-            className="w-40 sm:w-56 border-solid border-2 border-gray-200 rounded sm:ml-12 placeholder:text-xs indent-2.5 py-1"
+            className="input w-40 sm:w-56 sm:ml-12 placeholder:text-xs"
             placeholder="비밀번호를 입력해주세요"
             value={password}
             onChange={handleChange}
@@ -100,18 +100,10 @@ const LoginUser: React.FC = () => {
       <NaverLogin />
       <KakaoLogin />
       <div className="flex justify-between items-center w-full max-w-80 mb-12">
-        <button
-          type="button"
-          className="border-solid border-2 border-gray-200 rounded px-4 py-1 mt-12 hover:bg-gray-200 transition-all"
-          onClick={moveSignUpPage}
-        >
+        <button type="button" className="btn-blue w-24 mt-12 transition-all" onClick={moveSignUpPage}>
           회원가입
         </button>
-        <button
-          type="button"
-          className="border-solid border-2 border-gray-200 rounded px-4 py-1 mt-12 hover:bg-gray-200 transition-all"
-          onClick={() => navigate("/PasswordReset")}
-        >
+        <button type="button" className="btn-blue w-24 mt-12 transition-all" onClick={() => navigate("/PasswordReset")}>
           비밀번호 찾기
         </button>
       </div>
