+++ src/pages/Login.tsx
@@ -1,0 +1,11 @@
+import LoginUser from "../components/layout/login/LoginUser";
+
+const Login = (): JSX.Element => {
+  return (
+    <>
+      <LoginUser />
+    </>
+  );
+};
+
+export default Login;
