+++ src/router/Router.tsx
@@ -1,10 +1,10 @@
 import { Route, Routes } from "react-router";
-import Main from "../pages/MainPage";
-import Error from "../pages/ErrorPage";
+import Main from "../pages/Main";
+import Error from "../pages/Error";
 import ChannelPage from "../pages/ChannelPage";
 import ScheduleEditPage from "../pages/ScheduleEditPage";
-import LoginUser from "../pages/LoginUser";
-import SignUpUser from "../pages/SignUp";
+import Login from "../pages/Login";
+import SignUp from "../pages/SignUp";
 import PasswordReset from "../pages/PasswordReset";
 import KakaoLoginCallback from "../auth/kakao/KakaoLoginCallback";
 import NaverLoginCallback from "../auth/naver/NaverLoginCallback";
@@ -19,8 +19,8 @@ const Router = (): JSX.Element => {
     <Routes>
       <Route path="*" element={<Error />} />
       <Route path="/" element={<Main />} />
-      <Route path="/login" element={<LoginUser />} />
-      <Route path="/SignUp" element={<SignUpUser />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/SignUp" element={<SignUp />} />
       <Route path="/PasswordReset" element={<PasswordReset />} />
       <Route path="/auth/kakao/KakaoLoginCallback" element={<KakaoLoginCallback />} />
       <Route path="/auth/naver/NaverLoginCallback" element={<NaverLoginCallback />} />
